Move credential related functions to PostgresServerTreeItem
