+++ src/postgres/getPostgresCredentials.ts
@@ -6,15 +6,15 @@
 import { GenericTreeItem, IActionContext } from "vscode-azureextensionui";
 import { ext } from "../extensionVariables";
 import { nonNullProp } from "../utils/nonNull";
-import { PostgresDatabaseTreeItem } from "./tree/PostgresDatabaseTreeItem";
+import { PostgresServerTreeItem } from "./tree/PostgresServerTreeItem";
 
-export async function getPostgresCredentials(context: IActionContext, treeItem?: PostgresDatabaseTreeItem | GenericTreeItem): Promise<void> {
+export async function getPostgresCredentials(context: IActionContext, treeItem?: PostgresServerTreeItem | GenericTreeItem): Promise<void> {
     if (!treeItem) {
-        treeItem = await ext.tree.showTreeItemPicker(PostgresDatabaseTreeItem.contextValue, context);
+        treeItem = await ext.tree.showTreeItemPicker(PostgresServerTreeItem.contextValue, context);
     } else if (treeItem instanceof GenericTreeItem) {
-        treeItem = nonNullProp(treeItem, 'parent');
+        treeItem = nonNullProp(nonNullProp(treeItem, 'parent'), 'parent');
     }
 
-    await (<PostgresDatabaseTreeItem>treeItem).promptForCredentials();
+    await (<PostgresServerTreeItem>treeItem).promptForCredentials();
     await treeItem.refresh();
 }
