+++ src/postgres/tree/PostgresDatabaseTreeItem.ts
@@ -10,16 +10,11 @@ import * as vscode from 'vscode';
 import { AzExtTreeItem, AzureParentTreeItem, GenericTreeItem, IParsedError, ISubscriptionContext, parseError } from 'vscode-azureextensionui';
 import { getThemeAgnosticIconPath } from '../../constants';
 import { ext } from '../../extensionVariables';
-import { KeyTar, tryGetKeyTar } from '../../utils/keytar';
 import { localize } from '../../utils/localize';
 import { nonNullProp } from '../../utils/nonNull';
 import { PostgresSchemaTreeItem } from './PostgresSchemaTreeItem';
 import { PostgresServerTreeItem } from './PostgresServerTreeItem';
 
-interface IPersistedServer {
-    id: string;
-    username: string;
-}
 const invalidCredentialsErrorType: string = '28P01';
 
 export class PostgresDatabaseTreeItem extends AzureParentTreeItem<ISubscriptionContext> {
@@ -29,15 +24,9 @@ export class PostgresDatabaseTreeItem extends AzureParentTreeItem<ISubscriptionC
     public readonly databaseName: string;
     public readonly parent: PostgresServerTreeItem;
 
-    private readonly _serviceName: string = "ms-azuretools.vscode-cosmosdb.postgresPasswords";
-    private _keytar: KeyTar | undefined;
-    private _serverId: string;
-
     constructor(parent: PostgresServerTreeItem, databaseName: string) {
         super(parent);
         this.databaseName = databaseName;
-        this._keytar = tryGetKeyTar();
-        this._serverId = nonNullProp(this.parent.server, 'id');
     }
 
     public get label(): string {
@@ -57,7 +46,7 @@ export class PostgresDatabaseTreeItem extends AzureParentTreeItem<ISubscriptionC
     }
 
     public async loadMoreChildrenImpl(_clearCache: boolean): Promise<AzExtTreeItem[]> {
-        const { username, password } = await this.getCredentialsFromKeytar();
+        const { username, password } = await this.parent.getCredentialsFromKeytar();
 
         if (username && password) {
             try {
@@ -89,64 +78,6 @@ export class PostgresDatabaseTreeItem extends AzureParentTreeItem<ISubscriptionC
             label: localize('enterCredentials', 'Enter server credentials to connect to "{0}"...', this.parent.label),
             commandId: 'cosmosDB.getPostgresCredentials'
         })];
-    }
-
-    public async promptForCredentials(): Promise<{ username: string, password: string }> {
-        let username: string = await ext.ui.showInputBox({
-            prompt: localize('enterUsername', 'Enter username for server "{0}"', this.parent.label),
-            validateInput: (value: string) => { return (value && value.length) ? undefined : localize('usernameCannotBeEmpty', 'Username cannot be empty.'); }
-        });
-
-        const usernameSuffix: string = `@${this.parent.server.name}`;
-        if (!username.includes(usernameSuffix)) {
-            username += usernameSuffix;
-        }
-
-        const password: string = await ext.ui.showInputBox({
-            prompt: localize('enterPassword', 'Enter password for server "{0}"', this.parent.label),
-            password: true,
-            validateInput: (value: string) => { return (value && value.length) ? undefined : localize('passwordCannotBeEmpty', 'Password cannot be empty.'); }
-        });
-
-        await this.persistServer(username, password);
-        return { username, password };
-    }
-
-    private async getCredentialsFromKeytar(): Promise<{ username: string | undefined, password: string | undefined }> {
-        let username: string | undefined;
-        let password: string | undefined;
-
-        const storedValue: string | undefined = ext.context.globalState.get(this._serviceName);
-        if (storedValue && this._keytar) {
-            const servers: IPersistedServer[] = JSON.parse(storedValue);
-            for (const server of servers) {
-                if (server.id === this._serverId) {
-                    username = server.username;
-                    password = await this._keytar.getPassword(this._serviceName, this._serverId) || undefined;
-                    break;
-                }
-            }
-        }
-
-        return { username, password };
-    }
-
-    private async persistServer(username: string, password: string): Promise<void> {
-        if (this._keytar) {
-            const storedValue: string | undefined = ext.context.globalState.get(this._serviceName);
-            let servers: IPersistedServer[] = storedValue ? JSON.parse(storedValue) : [];
-
-            // Remove this server from the cache if it's there
-            servers = servers.filter((server: IPersistedServer) => { return server.id !== this._serverId; });
-
-            const newServer: IPersistedServer = {
-                id: this._serverId,
-                username
-            };
-            servers.push(newServer);
-            await ext.context.globalState.update(this._serviceName, JSON.stringify(servers));
-            await this._keytar.setPassword(this._serviceName, this._serverId, password);
-        }
     }
 }
