+++ src/postgres/tree/PostgresServerTreeItem.ts
@@ -8,11 +8,19 @@ import { DatabaseListResult, Server } from 'azure-arm-postgresql/lib/models';
 import * as vscode from 'vscode';
 import { AzExtTreeItem, AzureParentTreeItem, createAzureClient, ISubscriptionContext } from 'vscode-azureextensionui';
 import { getThemeAgnosticIconPath } from '../../constants';
+import { ext } from '../../extensionVariables';
 import { azureUtils } from '../../utils/azureUtils';
+import { KeyTar, tryGetKeyTar } from '../../utils/keytar';
+import { localize } from '../../utils/localize';
 import { nonNullProp } from '../../utils/nonNull';
 import { PostgresDatabaseTreeItem } from './PostgresDatabaseTreeItem';
 import { PostgresSchemaTreeItem } from './PostgresSchemaTreeItem';
 import { PostgresTableTreeItem } from './PostgresTableTreeItem';
+
+interface IPersistedServer {
+    id: string;
+    username: string;
+}
 
 export class PostgresServerTreeItem extends AzureParentTreeItem<ISubscriptionContext> {
     public static contextValue: string = "postgresServer";
@@ -20,9 +28,15 @@ export class PostgresServerTreeItem extends AzureParentTreeItem<ISubscriptionCon
     public readonly childTypeLabel: string = "Database";
     public readonly server: Server;
 
+    private readonly _serviceName: string = "ms-azuretools.vscode-cosmosdb.postgresPasswords";
+    private _keytar: KeyTar | undefined;
+    private _serverId: string;
+
     constructor(parent: AzureParentTreeItem, server: Server) {
         super(parent);
         this.server = server;
+        this._keytar = tryGetKeyTar();
+        this._serverId = nonNullProp(this.server, 'id');
     }
 
     public get iconPath(): string | vscode.Uri | { light: string | vscode.Uri; dark: string | vscode.Uri } {
@@ -71,6 +85,64 @@ export class PostgresServerTreeItem extends AzureParentTreeItem<ISubscriptionCon
                 return true;
             default:
                 return false;
+        }
+    }
+
+    public async promptForCredentials(): Promise<{ username: string, password: string }> {
+        let username: string = await ext.ui.showInputBox({
+            prompt: localize('enterUsername', 'Enter username for server "{0}"', this.label),
+            validateInput: (value: string) => { return (value && value.length) ? undefined : localize('usernameCannotBeEmpty', 'Username cannot be empty.'); }
+        });
+
+        const usernameSuffix: string = `@${this.server.name}`;
+        if (!username.includes(usernameSuffix)) {
+            username += usernameSuffix;
+        }
+
+        const password: string = await ext.ui.showInputBox({
+            prompt: localize('enterPassword', 'Enter password for server "{0}"', this.label),
+            password: true,
+            validateInput: (value: string) => { return (value && value.length) ? undefined : localize('passwordCannotBeEmpty', 'Password cannot be empty.'); }
+        });
+
+        await this.persistServer(username, password);
+        return { username, password };
+    }
+
+    public async getCredentialsFromKeytar(): Promise<{ username: string | undefined, password: string | undefined }> {
+        let username: string | undefined;
+        let password: string | undefined;
+
+        const storedValue: string | undefined = ext.context.globalState.get(this._serviceName);
+        if (storedValue && this._keytar) {
+            const servers: IPersistedServer[] = JSON.parse(storedValue);
+            for (const server of servers) {
+                if (server.id === this._serverId) {
+                    username = server.username;
+                    password = await this._keytar.getPassword(this._serviceName, this._serverId) || undefined;
+                    break;
+                }
+            }
+        }
+
+        return { username, password };
+    }
+
+    private async persistServer(username: string, password: string): Promise<void> {
+        if (this._keytar) {
+            const storedValue: string | undefined = ext.context.globalState.get(this._serviceName);
+            let servers: IPersistedServer[] = storedValue ? JSON.parse(storedValue) : [];
+
+            // Remove this server from the cache if it's there
+            servers = servers.filter((server: IPersistedServer) => { return server.id !== this._serverId; });
+
+            const newServer: IPersistedServer = {
+                id: this._serverId,
+                username
+            };
+            servers.push(newServer);
+            await ext.context.globalState.update(this._serviceName, JSON.stringify(servers));
+            await this._keytar.setPassword(this._serviceName, this._serverId, password);
         }
     }
 }
